INT-2915 JMS Remove Deprecated Attributes
delivery-mode was deprecated (in favor of delivery-persistent)
in 2.0. Remove attribute from schema and parser.

Unfortunately, the setter on the gateway was never deprecated
so there is a possibility someone is using it programmatically,
with no knowledge that it is going away.

Deprecated the setter.

Also, added schema documentation for several attributes where
no documentation was provided.