INT-2957 Remove Unnecessary AspectJ Dependencies
Core, JDBC, JMX, JPA each had dependencies for both
weaver and rt jars. weaver is a superset of rt.

Remove dependencies on the rt jar.

Further JDBC, JPA have no dependencies on AspectJ.

Finally, the JMX dependency was compile, when only
testCompile is needed.

However, to avoid breaking applications using 2.2.x
that might be relying on the transitive dependency
it is left at compile in this release.