INT-3053 Allow task-executor on <reply-listener/>
JMS Outbound Gateway did not allow the specification of
a task executor on the <reply-listener/> element.

This could prevent the gateway being used with a
listener in a JEE container.

Allow setting a task-executor.

Add a bean name to the reply listener container so
that a default TE will create suitably named threads.

INT-3053 Polishing

Fix container bean name when the gateway has no
bean name.