Update JavaMail dependencies
This commit updates the JavaMail dependency to use the new coordinates,
although the version that's being used remains unchanged. This brings
the dependency into line with other javax dependencies. It also paves
the way for moving to 1.5 at some point in the future as 1.5 no longer
publishes the javax.mail:mail artifact